+++ src/main/java/frc/taurus/logger/FlatBuffersLogger.java
@@ -60,7 +60,7 @@ public class FlatBuffersLogger {
       var channelType = channel.getNum();
       var reader = channelReaderMap.get(channel);
       while (!reader.isEmpty()) {
-        short queueSize = (short)reader.size();
+        byte queueSize = (byte)reader.size();
         ByteBuffer bb = reader.read().get(); // we know Optional::isPresent() is true because of earlier !isEmpty()
         writePacket(channelType, queueSize, bb); // write to file
       }
@@ -68,7 +68,7 @@ public class FlatBuffersLogger {
     // writer.flush();
   }
 
-  public void writePacket(final short channelType, final short queueSize, final ByteBuffer bbPayload) {
+  public void writePacket(final byte channelType, final byte queueSize, final ByteBuffer bbPayload) {
     int payloadSize = bbPayload.remaining();
 
     FlatBufferBuilder builder = new FlatBufferBuilder(maxHeaderSize + payloadSize);
